+++ include/kadnode/ext-bob.h
