+++ include/kadnode/ext-tls-server.h
@@ -2,7 +2,7 @@
 #ifndef _EXT_TLS_SERVER_H_
 #define _EXT_TLS_SERVER_H_
 
-#include "searches.h"
+#include <kadnode/searches.h>
 
 // Add domain via cert and key file
 int tls_server_add_sni( const char crt_file[], const char key_file[] );
