+++ include/kadnode/utils.h
