+++ src/announces.c
@@ -5,18 +5,18 @@
 #include <string.h>
 #include <limits.h>
 
-#include "log.h"
-#include "conf.h"
-#include "utils.h"
-#include "net.h"
-#include "kad.h"
+#include "kadnode/log.h"
+#include "kadnode/conf.h"
+#include "kadnode/utils.h"
+#include "kadnode/net.h"
+#include "kadnode/kad.h"
 #ifdef BOB
-#include "ext-bob.h"
+#include "kadnode/ext-bob.h"
 #endif
 #ifdef TLS
-#include "ext-tls-client.h"
+#include "kadnode/ext-tls-client.h"
 #endif
-#include "announces.h"
+#include "kadnode/announces.h"
 
 
 // Announce values every 20 minutes
