+++ src/ext-lpd.c
@@ -10,13 +10,13 @@
 #include <unistd.h>
 #include <assert.h>
 
-#include "main.h"
-#include "conf.h"
-#include "log.h"
-#include "utils.h"
-#include "net.h"
-#include "kad.h"
-#include "ext-lpd.h"
+#include "kadnode/main.h"
+#include "kadnode/conf.h"
+#include "kadnode/log.h"
+#include "kadnode/utils.h"
+#include "kadnode/net.h"
+#include "kadnode/kad.h"
+#include "kadnode/ext-lpd.h"
 
 
 enum {
