+++ src/ext-tls-server.c
@@ -20,14 +20,14 @@
 #include "mbedtls/oid.h"
 #include "mbedtls/error.h"
 
-#include "main.h"
-#include "conf.h"
-#include "log.h"
-#include "utils.h"
-#include "kad.h"
-#include "net.h"
-#include "searches.h"
-#include "ext-tls-server.h"
+#include "kadnode/main.h"
+#include "kadnode/conf.h"
+#include "kadnode/log.h"
+#include "kadnode/utils.h"
+#include "kadnode/kad.h"
+#include "kadnode/net.h"
+#include "kadnode/searches.h"
+#include "kadnode/ext-tls-server.h"
 
 
 /*
