+++ src/kadnode-ctl.c
@@ -11,7 +11,7 @@
 #include <arpa/inet.h>
 #include <netdb.h>
 
-#include "main.h"
+#include "kadnode/main.h"
 
 #define STR_HELPER(x) #x
 #define STR(x) STR_HELPER(x)
