+++ src/utils.c
@@ -10,10 +10,10 @@
 #include <netdb.h>
 #include <ctype.h>
 
-#include "main.h"
-#include "log.h"
-#include "conf.h"
-#include "utils.h"
+#include "kadnode/main.h"
+#include "kadnode/log.h"
+#include "kadnode/conf.h"
+#include "kadnode/utils.h"
 
 
 // Try to create a DHT id from a sanitized hex query
